usb: cdc_acm: Handle transfers larger than endpoint size
Using the usb_transfer API assumes that bulk transfers are never
going to be larger than wMaxPacketSize. This however does not have
to be the case.

For example Linux can send up to 20x larger transfer at once.

This patch uses endpoint events directly to process the data stream
as there is no need to segment the incoming data into individual
transfers anyway.

Signed-off-by: Josef Gajdusek <atx@atx.name>